PluginMaker overhaul to remove cached config state. Fixes #1532.
