+++ CHANGES.txt
@@ -6,6 +6,9 @@ Bug Handling
 
 * Correctly honor "user-agent" heading config in HttpInput (#1520).
 
+* Removed state from PluginMaker to remove race conditions during plugin
+  construction (#1532).
+
 0.9.2 (2015-04-22)
 ==================
