+++ plugins/process/process_directory_input.go
@@ -18,12 +18,13 @@ package process
 import (
 	"errors"
 	"fmt"
-	"github.com/bbangert/toml"
-	. "github.com/mozilla-services/heka/pipeline"
 	"os"
 	"path/filepath"
 	"strconv"
 	"strings"
+
+	"github.com/bbangert/toml"
+	. "github.com/mozilla-services/heka/pipeline"
 )
 
 type ProcessEntry struct {
@@ -201,16 +202,25 @@ func (pdi *ProcessDirectoryInput) procDirWalkFunc(path string, info os.FileInfo,
 		pdi.ir.LogError(fmt.Errorf("loading process file '%s': %s", path, err))
 		return nil
 	}
-	// Override the config settings we manage, make the runner, and store the entry.
-	entry.config.TickerInterval = uint(tickInterval)
 
-	commonConfig := entry.maker.CommonTypedConfig().(CommonInputConfig)
-	commonConfig.Retries = RetryOptions{
-		MaxDelay:   "30s",
-		Delay:      "250ms",
-		MaxRetries: -1,
+	// Override the config settings we manage, make the runner, and store the
+	// entry.
+	prepConfig := func() (interface{}, error) {
+		config, err := entry.maker.OrigPrepConfig()
+		if err != nil {
+			return nil, err
+		}
+		processInputConfig := config.(*ProcessInputConfig)
+		processInputConfig.TickerInterval = uint(tickInterval)
+		return processInputConfig, nil
 	}
-	entry.maker.SetCommonTypedConfig(commonConfig)
+	config, err := prepConfig()
+	if err != nil {
+		pdi.ir.LogError(fmt.Errorf("prepping config: %s", err.Error()))
+		return nil
+	}
+	entry.config = config.(*ProcessInputConfig)
+	entry.maker.SetPrepConfig(prepConfig)
 
 	runner, err := entry.maker.MakeRunner("")
 	if err != nil {
@@ -242,13 +252,24 @@ func (pdi *ProcessDirectoryInput) loadProcessFile(path string) (*ProcessEntry, e
 
 	mutMaker := maker.(MutableMaker)
 	mutMaker.SetName(path)
-	if err = mutMaker.PrepConfig(); err != nil {
-		return nil, fmt.Errorf("can't prep config: %s", err)
+
+	prepCommonTypedConfig := func() (interface{}, error) {
+		commonTypedConfig, err := mutMaker.OrigPrepCommonTypedConfig()
+		if err != nil {
+			return nil, err
+		}
+		commonInput := commonTypedConfig.(CommonInputConfig)
+		commonInput.Retries = RetryOptions{
+			MaxDelay:   "30s",
+			Delay:      "250ms",
+			MaxRetries: -1,
+		}
+		return commonInput, nil
 	}
+	mutMaker.SetPrepCommonTypedConfig(prepCommonTypedConfig)
 
 	entry := &ProcessEntry{
-		maker:  mutMaker,
-		config: mutMaker.Config().(*ProcessInputConfig),
+		maker: mutMaker,
 	}
 	return entry, nil
 }
